healthcare-cloud-system/microservices/notification-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
kafka-python==2.0.2
boto3==1.29.7
pydantic==2.5.0
prometheus-client==0.19.0
python-dotenv==1.0.0